+++ server-common/src/test/java/org/apache/gravitino/server/authorization/TestGravitinoAuthorizerProvider.java
@@ -33,7 +33,7 @@ public class TestGravitinoAuthorizerProvider {
     GravitinoAuthorizer gravitinoAuthorizer =
         GravitinoAuthorizerProvider.getInstance().getGravitinoAuthorizer();
     Assertions.assertInstanceOf(
-        IgnoreAuthorizer.class,
+        PassThroughAuthorizer.class,
         gravitinoAuthorizer,
         "Error initializing GravitinoAuthorizerProvider");
   }
